Changed list length to size_t and matched return type of utils_list_to_array.

+++ src/utils/list.c
@@ -31,7 +31,7 @@ utils_list_node_s* utils_list_del(utils_list_s *list, utils_list_node_s *node)
 	return node;
 }
 
-int utils_list_to_array(utils_list_s *list, utils_list_node_s **array)
+size_t utils_list_to_array(utils_list_s *list, utils_list_node_s **array)
 {
 	utils_list_node_s *itr;
